Move rename script from gulpfile.js to gulp-lib
